Tehtävät 4.15 - 4.23

+++ osa4/controllers/blogs.js
@@ -1,9 +1,18 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const middleware = require('../utils/middleware')
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const id = request.params.id
+  const blog = await Blog.findById(id)
+  const user = request.user
+
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(403).end()
+  }
 })
 
 blogsRouter.get('/:id', async (request, response) => {
@@ -17,40 +26,42 @@ blogsRouter.get('/', async (_request, response) => {
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   const body = request.body
-  if (body.title === undefined 
+  const user = request.user
+
+  if (body.title === undefined
     || body.url === undefined
     || body.title === ''
     || body.url === ''
-  ) 
-  {
+  ) {
     response.status(400).json()
   } else {
     const newBlog = {
       title: body.title,
       author: body.author,
       url: body.url,
-      likes: body.likes || 0
+      likes: body.likes || 0,
+      user: user._id
     }
 
     const blog = new Blog(newBlog)
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
     response.status(201).json(savedBlog)
-
   }
 
 })
 
-blogsRouter.put('/:id', (request, response, next) => {
+blogsRouter.put('/:id', middleware.userExtractor, (request, response, next) => {
   const body = request.body
 
-  if (body.title === undefined 
+  if (body.title === undefined
     || body.url === undefined
     || body.title === ''
     || body.url === ''
-  ) 
-  {
+  ) {
     response.status(400).json()
   } else {
     const blog = {
@@ -61,10 +72,10 @@ blogsRouter.put('/:id', (request, response, next) => {
     }
 
     const updatedBlog = Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
-    .then(updatedBlog => {
-      response.status(200).json(updatedBlog)
-    })
-    .catch(error => next(error))
+      .then(updatedBlog => {
+        response.status(200).json(updatedBlog)
+      })
+      .catch(error => next(error))
   }
 })
